app/requirements.txt
# FastAPI & Web
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9

# LLM
httpx==0.26.0

# Document Processing
pdfplumber==0.10.3
python-docx==1.1.0
pandas==2.2.0
openpyxl==3.1.2

# Utils
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
aiofiles==23.2.1
jieba==0.42.1
